elito-base-files: do not install qt.env when empty
other parts of the toolchain can create this file dynamically

+++ meta/recipes/elito-base-files/elito-base-files.bb
@@ -39,7 +39,9 @@ do_install() {
 		install -D -p -m 0644 $i ${D}${sysconfdir}/$i
 	done
 
-	install -D -m 0644 qt.env ${D}${sysconfdir}/default/qt.env
+	if test -s qt.env; then
+		install -D -m 0644 qt.env ${D}${sysconfdir}/default/qt.env
+	fi
 
 	for i in resolv.conf.static ntpd.conf.static; do
 		touch ${D}${sysconfdir}/$i
